Remove thread_task.h and thread_task.cc
