+++ c/parallel/parallel_for_dynamic.cc
@@ -16,7 +16,6 @@
 #include "parallel/api.h"
 #include "parallel/thread_pool.h"
 #include "parallel/thread_scheduler.h"
-#include "parallel/thread_task.h"
 #include "utils/assert.h"
 #include "utils/function.h"
 #include "utils/macros.h"          // cache_aligned
