--- c/parallel/thread_task.cc
@@ -1,32 +1,0 @@
-//------------------------------------------------------------------------------
-// Copyright 2019 H2O.ai
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//------------------------------------------------------------------------------
-#include "parallel/thread_task.h"
-#include "parallel/thread_worker.h"
-namespace dt {
-
-thread_task::~thread_task() {}
-
-
-
-simple_task::simple_task(function<void()> f_) : f(f_) {}
-
-void simple_task::execute(thread_worker*) {
-  f();
-}
-
-
-
-} // namespace dt
