--- c/parallel/thread_task.h
@@ -1,43 +1,0 @@
-//------------------------------------------------------------------------------
-// Copyright 2019 H2O.ai
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//------------------------------------------------------------------------------
-#ifndef dt_PARALLEL_THREAD_TASK_h
-#define dt_PARALLEL_THREAD_TASK_h
-#include "utils/function.h"
-namespace dt {
-using std::size_t;
-
-// forward-declare
-class thread_worker;
-
-
-class thread_task {
-  public:
-    virtual ~thread_task();
-    virtual void execute(thread_worker*) = 0;
-};
-
-
-class simple_task : public thread_task {
-  private:
-    function<void()> f;
-  public:
-    simple_task(function<void()>);
-    void execute(thread_worker*) override;
-};
-
-
-} // namespace dt
-#endif
